尝试修复编译错误

+++ scripts/clearDist.js
@@ -1,10 +1,13 @@
 const path = require("path");
+const fs = require("fs");
 const rimraf = require("rimraf");
 
 const distPath = path.resolve(__dirname, "../dist");
 
 async function run() {
-    await rimraf(distPath);
+    if (fs.existsSync(distPath)) {
+        await rimraf(distPath);
+    }
 }
 
 run();
